Create Geofs Carrier Cats
This is the Carrier Catapult in GeoFS. Press 'q' to lower the launch bar, then line up with a catapult and press '/' to lock it in place. Throttle max and press 'l' to launch. Aircraft supported: F-16, F-14, and E-2C. Please check this creator out. His name is @Ariakim-Taiyo .